requirements.txt
aiohttp==3.9.5
aiohttp-cors==0.7.0
aiosignal==1.3.1
attrs==23.2.0
customtkinter==5.2.2
darkdetect==0.8.0
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
numpy==1.26.4
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
websockets==12.0
yarl==1.9.4
